add middleware
The middleware will check for if SECRET_KEY is set to something other
than 'UNSAFE_DEFAULT' and if it is, and if that secret key is not passed
in a get/post request with the param name 'secret', it returns a 403
error.  If SECRET_KEY is set to 'UNSAFE_DEFAULT' or if the secret key
value is passed as get/post param 'secret', it returns None, which
allows requests to graphite to proceed.

The value of SECRET_KEY will either be imported from
graphite.local_auth.secret or will be set to the fallback value of
'UNSAFE_DEFAULT'.  This allows the creation of a volume on container
start (using the '-v' flag on 'docker run') at the container directory
/opt/graphite/webapp/graphite/local_auth that is mapped to some
directory on the host machine containing an __init__.py where SECRET_KEY
can be specified.